stage1: prepare-app: don't mount /sys if path already used
When users mount /sys or a sub-directory of /sys as a volume,
prepare-app should not mount /sys: that would mask the volume provided
by users. This patch detects if there is a volume mounted in /sys or a
subdirectory and skips the mount.

Fixes https://github.com/coreos/rkt/issues/2874