rkt/export: refuse to export if the app has MPs + no-overlay
If we're using "fly" the mounts we do (`/dev`, `/proc`, `/sys` and
`/tmp` and other user-defined mounts) are in the host's mount namespace.
Trying to export this kind of pods causes problems (e.g. some files in
`/proc` cannot be directly read). Also, the user doesn't expect the
contents of mountpoints to be exported.

This is not problematic if we use overlay, because in that case we
derive the exported fs from the ACI image plus the upper layer.

To avoid problems, if the pod doesn't use overlay, we refuse exporting
if the app has mountpoints inside.