scripts: add /etc/rkt owned by group rkt-admin in setup-data-dir.sh
Fixes the immediate problem with `scripts/setup-data-dir.sh` being out
of sync with `init/systemd/tmpfiles.d/rkt.conf`(#2862).

This doesn't include the tips mentioned by @lucab in the issue. I'm
working on it. Just wanted to take a first stab to fix the problem for
now while I work on the rest of the points mentioned in #2862 to make
the script better and avoid duplicating the directories.

Signed-off-by: Akshay Karle <akshay.a.karle@gmail.com>