store: allow users in the rkt group to delete images
Don't check if we're the owner of the image before removing, all the
users in the rkt group should be able to remove any image.

If the user removing the image is in the rkt group and permissions are
correctly set up (or if they're root), then it's fine, they can remove
images.

If the user is not in the rkt group or permissions are not correctly set
up, they will get errors before this function because they're not
allowed to use the CAS directory.