pkg/tar/chroot: avoid errwrap in function called by multicall
When a function is called by multicall, it is not possible to use errwrap
because the log pkg is not initialized yet. Also initialize the log pkg will
not show the correct indentation in the log because the new process will
start printing from the most left.

Fixes https://github.com/coreos/rkt/issues/2984