functional tests: skip test if remote fetch fails
On eed65378a we started skipping tests that fail on remote fetch when
calling rkt run or prepare. However, we didn't do the same with the
fetching functions we run before those.

Modify those functions to return an error if the exit status is not 0
and skip the test when this happens.

On the other tests, we always fetch from local files so treat errors as
Failures.