trust: provide InsecureSkipTLSCheck to pubkey manager
We should be able to pass --insecure-options=tls option to the `rkt trust`
command. But now we can't because we test insecure tls value not from
global options but from already initialized public key manager in the
metaDiscoverPubKeyLocations():

if m.InsecureSkipTLSCheck {
        insecure = insecure | discovery.InsecureTLS
}

This patch adds InsecureSkipTLSCheck to the initialization of a public
key manager structure during `rkt trust` command execution. So, the
value of the `--insecure-option=tls` will be matched and used as we
suppose.