stage1/init,prepare_app: allow devices in volumes
Use `DeviceAllow=` to allow accessing to devices that are in volumes.

Since systemd requires the device to start with `/dev` we need to do
some tricks to make it work.

The device won't exist inside the container so for a given mount, if the
volume is a device node, we create a symlink to the mount target in
`/rkt/volumes`. Later, prepare-app copies the links in `/rkt/volumes` to
`/dev/.rkt/` which is what we write in `DeviceAllow=`