build: Fix loading libnss
I could not find where systemd tries to load libnss_files, so probably
it does not do it directly. But apparently the loader does not like
the hardcoded /usr/lib path in the ACI rootfs, so instead put the
libnss_files libraries in the same directory as found on host.