build: Fix make clean for recent systemd version in src flavor
Apparently systemd got some tools written in python, which is executed
at the build time. And python being python likes to put the compiled
bytecode of the script next to the script itself, which violates my
assumption that the srcdir should not change when we perform the build
in a different directory.

To fix this, we generate the second filelist of the srcdir after the
building of systemd finishes and generate a clean file based on the
generated filelist.