rkt: Introduce Distribution Point concept
A Distribution Point represents the method to fetch an image starting from an
input string.

Before this concept was introduced, rkt used the ImageType concept, the
ImageType was mapped to the input image string format and internally
covered multiple concepts like distribution, transport and image type
(hidden since all are now appc ACIs).

The distribution point concept will be used as the primary information in the
different layers of rkt (fetching but also for references in a CAS/ref store).

This patch introduces the distribution point concept. See
Documentation/devel/distribution_point.md for a detailed description.