scripts: Fix missing files in .deb when using install-rkt.sh
Checkinstall (or rather installwatch) hooks into the mkdir process of
install and prefixes the given paths while creating them. Since mkdir
does a chdir($dir) after the process of creating the named directories
it detects, that the creation must have failed and bails. The
install-pak script continues with the next file in the list which is
placed in the fake directory as expected and doesn't fail since the
directory exists in the prefixed dir. Thus the first file in the loop is
missing from the prefix dir.
This is fixed with creating each directory that is required before the
normal install routine continues.

Fixes #3123